add schema comments and update file/document; unify VerticalAlign type
