--- src/file/document/body/section-properties/vertical-align/index.ts
@@ -1,2 +1,0 @@
-export * from "./vertical-align";
-export * from "./vertical-align-attributes";
