--- src/file/document/body/section-properties/vertical-align/vertical-align-attributes.ts
@@ -1,10 +1,0 @@
-import { XmlAttributeComponent } from "file/xml-components";
-import { SectionVerticalAlignValue } from "./vertical-align";
-
-export class SectionVerticalAlignAttributes extends XmlAttributeComponent<{
-    readonly verticalAlign?: SectionVerticalAlignValue;
-}> {
-    protected readonly xmlKeys = {
-        verticalAlign: "w:val",
-    };
-}
