+++ src/file/vertical-align/index.ts
@@ -1,0 +1,1 @@
+export * from "./vertical-align";
